services/notification_service/requirements.txt
Flask==2.3.2
Flask-Cors==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
mysql-connector-python==8.1.0
bcrypt==4.1.0
Flask-Mail==0.10.0
requests==2.31.0